add inchlib library to gene ex component
WIP sort of working, fixing javascript

EOD WIP - it's working
