+++ public/index.html
@@ -82,11 +82,6 @@
     <script src="/gdc-app-modified.js"></script>
     <!-- END legacy Angular stuff -->
 
-    <!-- inchlib stuff -->
-    <script src="konva.min.js" type="text/javascript"></script>
-    <script src="inchlib-1.2.0.js" type="text/javascript"></script>
-    <!-- end inchlib stuff -->
-
     <script
       language="JavaScript"
       type="text/javascript"
